@@ -1,13 +1,13 @@
 // tslint:disable-next-line:no-import-side-effect
-import { WlConfig } from "wl-range";
+import { WlConfig } from 'wl-range';
 
-import { applyPolyfills, defineCustomElements } from "wl-range/loader";
-import { WlWindow } from "./interfaces";
+import { applyPolyfills, defineCustomElements } from 'wl-range/loader';
+import { WlWindow } from './interfaces';
 
 export function appInitialize(config?: WlConfig) {
   const win: WlWindow = window as any;
-  const Ionic = (win.Wl = win.Wl || {});
+  const Wl = (win.Wl = win.Wl || {});
 
-  Ionic.config = config;
+  Wl.config = config;
   applyPolyfills().then(() => defineCustomElements(win));
 }
