@@ -1,23 +1,23 @@
-import { PluginFunction, default as VueImport } from "vue";
+import { App, Plugin } from 'vue';
 
-import { WlConfig } from "wl-range";
-import { appInitialize } from "./app-initialize";
+import { WlConfig } from 'wl-range';
+import { appInitialize } from './app-initialize';
 
-import { createInputComponent } from "./components/inputs";
+import { createInputComponent } from './components/inputs';
 
-let Vue: typeof VueImport;
+let Vue: App;
 
-export const install: PluginFunction<WlConfig> = (_Vue, config) => {
+export const install: Plugin = (_Vue, config: WlConfig) => {
   if (Vue && _Vue === Vue) {
-    if (process.env.NODE_ENV !== "production") {
-      console.error("[Wl] already installed. Vue.use(Wl) should be called only once.");
+    if (process.env.NODE_ENV !== 'production') {
+      console.error('[Wl] already installed. Vue.use(Wl) should be called only once.');
     }
     return;
   }
   Vue = _Vue;
-  Vue.config.ignoredElements.push(/^wl-/);
+  Vue.config.isCustomElement = tag => tag.startsWith('wl-');
 
-  createInputComponent("WlRangeVue", "wl-range");
+  createInputComponent('WlRangeVue', 'wl-range');
 
   appInitialize(config);
 };
